Change path of dev-tool
