+++ src/pages/dev-tool/index.module.scss
@@ -1,0 +1,54 @@
+@import "../../styles/variables";
+@import "../../styles/mixins";
+
+.wrapper {
+  @include mainArea;
+}
+
+.title {
+  @include pageTitle;
+}
+
+.panelWrapper {
+  margin-bottom: 16px;
+
+  &:hover {
+    .linkArrowIcon {
+      transform: translateX(8px);
+    }
+  }
+}
+
+.panelInner {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.panelContentRight {
+  display: flex;
+  align-items: center;
+}
+
+.panelIconBg {
+  background-color: $color-brand;
+  padding: 8px;
+  border-radius: 10px;
+  transition: transform 200ms ease-in-out;
+}
+
+.panelIcon {
+  font-size: 2em;
+  color: white;
+}
+
+.panelTitle {
+  margin-left: 16px;
+  font-size: 1.2em;
+}
+
+.linkArrowIcon {
+  font-size: 2em;
+  color: $color-gray-400;
+  transition: transform 200ms ease-in-out;
+}
